+++ zebra-script/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.0.0] - 2026-06-02
+
+### Changed
+
+- Update to `zebra-chain` 9.0.0 (NU6.2 support). No other changes to this crate.
+
 ## [7.0.1] - 2026-05-29
 
 This release fixes one consensus security issue: